Add is_oda attribute to reports
We have decided to use the same attribute name and the same conventions
for reports that we do for activities, in order to add the least amount
of confusion and friction to development.

The conventions are:

- For "pure ODA" funds such as GCRF, Newton, and Other ODA, the value of
`is_oda` is `nil`; this is due to the fact that these funds predate the
existence of hybrid (ODA and non-ODA) funds such as ISPF; these funds
and all their activities are *implicitly* ODA.

At the time of adding ISPF to RODA, we had decided against backfilling
the value of `is_oda` for existing activities. This allows us to enforce
the presence of `is_oda` for ISPF activities.

We acknowledge the tension that comes from having a boolean attribute
whose `nil` value does not translate to a "falsey" meaning in the
business domain. However, many RODA workflows now depend on this meaning
for decision paths.

- For hybrid funds such as ISPF, `is_oda` must be intentionally set at
programme level (level B), and is required; projects and third-party
projects (levels C and D) inherit the `is_oda` value of their parent
activities. Only `is_oda: true` means that the activity is ODA, and only
`is_oda: false` means that the activity is non-ODA.

We are populating the `is_oda` value for all existing ISPF reports to
`false`. We have confirmed with DSIT that all existing ISPF reports are
for non-ODA activities.