Editable report for activity has same ODA type
Most of the actions performed on projects and third-party projects
require the existence of an editable report corresponding to that
activity, so all the actions (adding or updating activities, adding
financial data, etc) can be reported on.

For ISPF reports, we need to update the query that retrieves the report
to take into account the ODA type (`is_oda`) of the activity.