Update projects for report finder
This service is responsible for retrieving all the projects and third-
party projects that _could_ have reportable actions for a given report.
Before introducing the concept that reports can be ODA or non-ODA, the
finder only took into account the fund and the organisation.

Since reports can now be ODA or non-ODA, the set of relevant activities
needs to be restricted by ODA type as well.