Pre-existing later report validation considers ODA
The check for a pre-existing later report is ensuring that no one is
rewriting history. Any data up to the latest report is considered to
have been already reported, so we don't want anyone to create a report
for a period preceding the latest report.

Previously, this check applied to the combination of fund and partner
organisation.

Now that ISPF reports can be for ODA or non-ODA, the check needs to take
into account the ODA type of the report, because the reporting streams
are separate and could function on different timelines.