Order reports by fund, ODA status, and FY/FQ
As part of adding the ODA/non-ODA information to report tables, we held
a design collaboration to determine what other changes DSIT users would
find the most helpful for these views.

Users would like reports (within their partner organisation group)
ordered by fund (with ODA coming before non-ODA for ISPF), as well as by
financial year and quarter, with the most recent at the top.

The desired order of the funds is not alphabetical. Instead, it happens
to coincide with the order in which the funds were added to RODA. We are
therefore using the fund's creation date as a proxy for the intended
order. (Adding a dedicated attribute by which to sort funds in the views
would take longer and delay other work.)