Show environment for the newly added dev domain
All non-production environments show a banner in the top left corner of
the page, and a prefix in emails sent from the application, to indicate
what environment they are coming from.

This is in order to avoid user confusion about what site they are on,
and reassure them that they are not accidentally modifying production
data without intending to.

We cannot use Rails.env, because the Rails env is production for all -
in that they are live sites, not local dev or testing. We have to use
the hostname from the `DOMAIN` environment variable.

When we originally added the banner, there was no dev environment, and
the script defaults to not showing a banner for any unrecognised hosts,
to avoid showing a banner unintentionally.

The pentest domain has been deprecated, so let's remove it to reduce
confusion. (It's possible that the same happened for sandbox, but let's
handle one rabbit hole at a time.)

We've also:
- reordered the env names alphabetically wherever they appear
as a list
- added "dev" to the domains that show an amber warning when logged into
a Rails console in non-sandboxed mode