Apply ODA eligibility criteria within report
Applying the ODA eligibility criteria globally, by including it in the
definition of `reportable` activities, is not correct in the global
context: ISPF non-ODA activities are reportable, as long as they have
the relevant `programme_status`.

Instead, we will apply the ODA eligibility criteria on a report basis,
as the report defines what it means for activities to be reportable
within its purview.

For example, only projects and third-party projects are currently
relevant for reporting, and it is the report's responsiblity to invoke
the service object that applies the activity level criteria.

(This is not to say that we couldn't rethink these responsibilities, but
in the current state, it's not much of a stretch to move this criterion
under the report.)