Update Node installation
The nodesource.com setup scripts are deprecated.

Here we update the Dockerfile to use the new recommended installation.

We've moved both Node and Yarn installation into APT as we consider
them 'system' layer tools which install the packages on which the
application depends.

We've exposed the `NODE_MAJOR` version as a build argument which
defaults to 16, the current version in use by the application.