Report presenter can present is oda
We want to show the value of the `is_oda` column on the report show
view.

To do so we extend the report presenter to handle presenting the value.

Right now we have to guard against a `nil` value as for all reports that
are not Internation Science Partnership Fund (ISPF) this value must be
`nil`, returning nil feels like the best thing to do there.

We have added a locale string to the very top of the hierachy as 'is
oda' is a concept shared across models.