Full fund name and ODA type on report edit form
When a report is for the ISPF fund it can be either ODA or non-ODA, we
want to show this on the report edit view appended to the full fund
name.

We have already allowed the report presenter to do this, here we show
the value.

When the report is not for the ISPF fund, we show only the name, as ODA
is ALWAYS assumed for all other funds.