Add total actuals column to report csv
The total actuals is the sum of the 'actuals net' columns in the report,
of which there can be multiple..

Here we keep a total of the actual net values as it is calculated for each
financial quarter and add it to the columns for the current row if
appropriate.

Because the ActivityActualsColumns class also provides data for the
'spending breakdown' we have to check that a report is being generated
and only include the total actuals column if that is the case.

Most changes here relate to the addition of the extra column, the
ActivityActualsColumns adds `n` number of financial quarters and then
adds the total actuals to the end, some of this code has to rely on the
position of the values i.e the column number which has been altered by
this work.

We could not see another approach for this other than a major refactor
of how the exports are put together, so we settled on making this work.