Import gets ODA type from report where applicable
Level C and D imports are always associated with a report, so they can
use the report's `is_oda` value instead of receiving it from the
controller.

We cannot remove the `is_oda` argument entirely, because the import
service is also used for level B activities, for which there is no
report, so the import still needs to receive the value of is_oda from
outside.