Fix the coverage calculation
We use Simplecov as part of Coveralls to generate a code coverage
percentage and a bar which we hold ourselves to (98%).

https://coveralls.io/
https://github.com/simplecov-ruby/simplecov

Unfortunately, Simplecov looks to mis-configured so our coverage report
may not have been as accurate as we would have liked!

By configuring Simplecov explicitally for Rails, we get a more accurate
report (the spec files themselves for example are no longer included
which were heavily skewing our result).

In order to continue to reach our goal of 98% coverage we have more
work to do or some decisions to make:

For now we are ignoring low coverage in some specific files and all the
tasks, these are common items to skip, but we should make a decision
about where to take these.

We've heavily commented the filters to account for this.

With this work in place we can be more confident about out coverage and
so our test suite.