Refactor export/report sepc
Having spent the day working with this spec we felt this refactor would
help.

When testing the values in a row we have to assume certain values are at
certain indexes in the array. Previously we tried to encode this with
methods like `roda_identifier_value_for_row` where the assumption is that
the RODA identifier is always at index 0.

When adding columns this means everything gets moved about, it is also
not clear to the reader what the index means without digging into the
headers to get the context.

We've replaced those methods with a single one that takes the column
header string and returns the value, this ties the index to the
'purpose' in a more understandable way and removes the need to interpret
the headers. This also ensures that the expected value is in the
expected column, lessening the brittlness slightly.

Whilst it is true that we do not test that the right column is at the
expected index, we did not do this previously so nothing is lost, whether
we should test that is a seperate issue!