Remove coveralls and update simplecov gems
We use coveralls to montior test coverage:

https://coveralls.io/github/UKGovernmentBEIS/beis-report-official-development-assistance

However our implemenation is out-of-date and no longer works as
expected.

To fix this we first remove the coveralls gem as it is no longer
required and add a newer version of simplcov (on which coveralls
depends).

We also add the simplecov-lcov formatter, this will be used later to
provide data to coveralls.

We retain the existing html formatter for local development.