Add support for Coveralls GitHub action
To get our coverage into Coveralls we have to make sure the coverage
report output is available in the runner.

To do this we first ensure the `coverage` directory is available on the
runner, as it is not commited.

We then 'bind' mount the directory directory into the container so that
when the tests run and the coverage report is written, it is also written
to the runner filesytem.

With the coverage report available, we can run the Coveralls action to
send the result to Coveralls.

We set `fail-on-error` so that CI still passes if the data cannot be
sent to Coveralls, our CI still fails if our coverage drops below the
required 98%.