Allow specs to skip AR primary key validation
Factory-bot-rails 6.4 has added a configuration setting and validation
that results in stricter checks for associations between persisted and
stubbed records:
https://github.com/thoughtbot/factory_bot_rails/commit/0040292890541736d0065c30ab2f798e8608787d#diff-d38ec77f185082fa818f22bea77b9c3597d01e783ea5a799469002738d739477R6

The intention as far as I can tell was to reject, by default, assigning
primary keys to factory-created objects. Our specs don't assign primary
keys to objects, but they associate stubbed objects to persisted objects
in some cases, and it's this that triggers the validation added by
factory-bot-rails. The stubbed object hasn't been persisted, so its id
fails validation when the spec tries to associate it with a persisted
object.

We want to keep our test suite from getting any slower, so we want to
use stubbed objects wherever possible.

The suggested fix was to add an explicit configuration on our side to
turn off `reject_primary_key_attributes`, which in turn means that the
active_record_validator will not be triggered.