Add IATI scope service
This service object takes an array of benefitting country codes and
returns the IATI 'scope' code, see:

https://iatistandard.org/en/iati-standard/203/codelists/activityscope/

Not all codes are covered by the data in the application, the logic used
was defined by the client:

- national (code 4)  when there is only a single benfitting country
- regional (code 2) when there are multiple benefitting countires within a single
  region
- multi-national (code 3) when there are multiple benefitting countires across
  multiple regions

Some activities do not have benefitting countries so we try to guard
against this, returning false so that calling code can take appropriate
action.

The main decisions we had to make in this work:

What approach? This is a big codebase with a long history, it can be
hard to find a consitent approach. We ened up settling on a small,
isolated service object as there are already many examples of this
pattern in the code.

The alternatives might include in the Activity class, but this is
already very large or in a module, which tends to 'hide' the code.

The service object has many benefits; It is a well trodden path, the
purpose of the pattern is to model business logic and it isolates that
logic for testing as well as leaving clear documentation on the business
logic used.

The other tough choice was what to return if no scope cannot be calculated,
we felt the choices were nil, false or to raise an error.

Given the purpose of the object we felt false was a good choice as we
only want to show the value if it can be calculated, using false allow
for the calling code to make a choice about what to do when the scope
cannot be calculated.