Remove specs for ISPF in stealth mode
The feature flag allowing everything to do with ISPF to be hidden has
been removed in https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/2192

This spec has escaped our notice because it was not using the feature
flag by name. Instead, it relied on a proxy for the flag being on or off
in the form of the funds being loaded in the exports controller as
`Fund.all` or `Fund.not_ispf`.

Since the flag has been removed, there are no conditions under which the
funds being loaded on the exports page are restricted to non-ISPF only.