Remove unused scopes and methods
Scopes and methods to return all but ISPF funds, activities, and reports
are now unused. They were only in use when the app had a feature flag to
enable hiding everything to do with ISPF.

I did hesitate for a bit about keeping this code "just in case", but
ultimately we know that in general, unused code is a recipe for
confusion and an increased burden of maintenance, and it can always be
added back if it's needed in future.