Fix flaky sign in spec
We have seen this spec fail in CI a number of times with what appered to
be the OTP sent to Notify no being the same as `otp_at_time_of_login`,
here is an example:

https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/actions/runs/7020362563/job/19099874908#step:6:9200

Our hypothesis is that this was caused by some kind of variance in the
time between the setting of `otp_at_time_of_login` and the expectation.

During this work we tried to understand how we have implemented the 2FA
in the application. This work seems to indicate that although we configure
the expiry and drift (a window of time to allow for differences in the
server and client time) of the token to 5 minutes, the acutal time is 5
minutes and 30 seconds.

Regardless, we decide the best we could do is use more strict time
control around when the user is signing in and when they enter the token
in this feature spec.

By using as close to a resolution in seconds (time is never going to be
perfect!) we can demonstrate both the expected behaviours:

- token is valid up to 5 minutes 29 seconds after creating it
- token is invalid from 5 minutes 30 seconds after creating it

To be honest, we are not sure at this point if this will prevent the
spec flaking, but it is a good start.

We will document what we've learnt in a later commit.