Sign in spec OTP is always correct
In previous work we tightened up the time control in this spec as we
were seeing occasional failures.

However, in the context of this spec, I don't think we care if the OTP token is
correct, only that:

- the right token is sent to Notify
- that with the correct token, the user is signed in

Here we have made some changes to the spec so that ANY token is
considered correct, this should decouple any time issues that relate to
the token.

We have left the time controls in as documentation for the expiring of
the token, even though they are not required.