Order activity comments exported in budget CSV
PostgreSQL doesn't guarantee an order for records unless we specify an
ordering criterion ourselves. It is *likely* that the records will be
returned with the most recently updated last, which is why the spec only
fails occasionally.

This is potentially the cause of the budget spec failing sometimes.

Specifying the order in which the comments are returned will fix the
intermittent failures, if they were caused by the unpredictable ordering
coming from the database.

If, on the other hand, the intermittent errors are caused by some
randomness in how the test objects are created[1], we may see failures
again sometimes.

[1] `programme_activity_comment_1` should be created before
`programme_activity_comment_2`. If *this* ordering is not guaranteed,
and `programme_activity_comment_2` ends up with a `created_at` older
than `programme_activity_comment_1`'s created_at, then the fix may not
make a difference.