Alias health_check so it responds to health-check
The AWS health check endpoint has been configured to look for the route
`health-check` instead of `health_check`.

Rather than asking them to reconfigure it, we will accept that both
versions (with dash and with underscore) are common, and provide an
alias for the route, so that it responds to both requests.