Add Actial and Refund csv import row
We want to offer users the ability to report:

- 0 actual
- 0 refund
- provide a comment as to the zeros

The existing 'actuals importer' handles both Actual and Refund
Transactions and has been extended over time.

The import process is quite complex as there are dependencies across the
columns.

Rather than continue to extend the existing class, we decided to create
a new one that will make the code more approachable and support the new
behaviour.

As the existing class is broken down over three loose classes with
responsibilities to:

- convert the supplied csv data to useable values
- validate the values
- authorise the activities
- create the actual objects

The responsibility is shared across the classes and can be hard to
follow. We decided that the shape of this problem has a lot of
similarities to accepting user supplied values from a web form and in
those instances, we will often reach for the 'form object' pattern, a
standin for the object we wish to create that can be validated
seperately and then used to create the model or models.

We felt this is pretty much the same thing here except the input is not
a set of strings submitted via a form, rahter strings from a CSV file.

The new `Import::Transactions::ActualsAndRefundCsvRow` class does just
this, for a single row in the CSV file:

- it returns any value from the csv that is required to import the
  modelled row
- it explicitally treats an 'empty' value as a 'blank'
- it validates the values seperately
- it retains the original values even when invalid to assit the user
  later
- it loosely behaves like an ActiveRecord model

The downside is that we have to duplicate all the validation from the
underlying models, but that price feels worth paying to make the code
more approachable - the result of calling `valid?` is an object than can
be used to create the underlying objects in the database or not, with
errors that explain why (just like feedback on a form!).

By loosely maintain the object interface, refactoring the actual
importing process is relatively simple.