Model financial values in csv for importing
To try and simplify our approach to importing from csv files this
Import::Csv::Financial class is a basic wrapper for a financial value
that is to be imported from csv.

Our philosophy when importing from csv is that a financial value is to expect:

- a positive number
- a negatvie number
- zero
- any kind of blank space
- invalid
- some kind of financial formatting e.g. 10,000.00

Here, invalid means

> We could not convert the supplied string to a BigDecimal.

This implementation treats a `0` and an empty cell the same,
which makes the upstream import code more approachable as we can
assume everything will be `0`.

We resue the existing `ConvertFinancialValue` service object to do the
actual conversion, dealing with an errors as described above - we chose
this approach over changing the `ConvertFinancialValue` to keep the
changes isolated to the current work.

When we use this class during import, we can expect a financial value to
be one of three states:

- a BigDecimal value
- a zero
- nil

Where nil means it could not be converted.

We also retain the original string value from the csv so it can be
presented back to the user in any error messages.