Model a actual, refund and comment upload row
We've already added the `Import::Csv:Financial` class for handling the
financial values from csv, actuals and refunds.

This commit adds a representation of the row as we expect it to be
supplied in the csv, we perform a set of validations adding to the
errors instance variable for use later.

Like financial values, we treat 'empty' values such as spaces as nothing
rather than included values that users would not see.

At this point a valid `Import::ActivityActualRefundComment::Row` still
may not result in a record being created as validation occurs later and
is collected together for presentation to the user.

We are doing more validation than the existing improter does, earlier, we
feel like this approach results in a number of benefits including:

- the code is more approachable, with as many validation happening as
  early as possible in one place
- performant, we don't have to waste resources during the import if we
  already know the row is invalid
- we can offer more helpful errors than the standard model validation might.

As part of this work we are introducing some better name and namespacing
for the code that relates to the importing processes. For related
models, services and any other we can use:

`import::csv:activity_refund_comment`

We settled on this as the code relates so many models, including:

- Actual Transactions
- Refund Transactions
- Comments on Activities

This model reflects the change to the expectation on users around what
values should be included in a csv row once it is in use.