Activity actual, refund and comment row import service
The responsibility here is to take a CSV::Row, report and user and create the
appropriate records based on the values in the row.

It relies on the `Import::Csv::ActivityActualRefundComment::Row` to do
it's work.

The result of running the service is either:

- the object that was created, a Actual, Refund or Comment record
- false

When false, `errors` will be populated with the collected errors for the
row. These errors include:

- the validation errors from `Import::Csv::ActivityActualRefundComment::Row`
- the validation errors form the attempt at creatng the object
- further validation on the row:
  - the Activity exists
  - the Activity can be modified by the user
  - the Acitivity is in the correct report

and will eventually be presented back to the user.

Later work will take a complete csv file and use this class to create
the objects, but we want to seperate this out for testing and
approachability.