Extend the RowError class
We want to reuse this error class for the activity, refund and comment
import process, but to do so it needs to support the existing interface
so that we can be compatible whilst the importer is switched to the new
one that is in progress.

All we need is a representation of the row as it appears in the csv data
file, the users expectation of the row number.

This is just a matter of the enumerable index plus 2.

Adding this method has no impact on existing code.

We also add a test.