Model a skipped row
When importing actual, refunds and comments from csv we now 'skip' empty
rows, these are rows that could be imported but do not contain data to
satisfy the importer.

Here we add a model for a row that meets this criteria that the
`Import::Csv::ActivityActualRefundComment::RowService` can return.

Doing to we let use show the user the rows in their data that were
skipped and verify that this was expected.