Fix create actual spec for factory bot rails
The change introduced in factory bot rails 6.4.3 impacts this spec.

https://github.com/thoughtbot/factory_bot_rails/commit/0040292890541736d0065c30ab2f798e8608787d#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R168

We can resolve this by stubbing the call to `organisation` on the
(stubbed) Activity, rather than trying to update it.

In the `does set the report` spec we have to work a bit harder if we
want to rely on the stubbed objects. Inside the CreateActual object we
call `Report.editable_for_activity` so we are relying on the database
and cannot rely on the stubbed objects (since factory bot rails 6.4.3).

However, the `editable_for_activity` is not under test here so we can
stubb that to always return the report which gives is the coverage we
need.