Fix create user spec for factory bot rails
The change introduced in factory bot rails 6.4.3 impacts this spec.

https://github.com/thoughtbot/factory_bot_rails/commit/0040292890541736d0065c30ab2f798e8608787d#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R168

We can resolve this here by simply not trying to pass a stubbed
organisation as passing an organisation and observing the result is
covered in a later spec, in the first three specs the organisation only
need to exists we don't really mind what it is.