Fix external income policy spec for factory bot rails
The change introduced in factory bot rails 6.4.3 impacts this spec.

https://github.com/thoughtbot/factory_bot_rails/commit/0040292890541736d0065c30ab2f798e8608787d#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R168

One we switch to Factory bot Rails > 6.4.3 we began to see Bullet
errors when running this spec.

The issue here appears to be caused by `build_stubbed` objects no longer
behaving like database objects (which they should not have been in the
first place), with the most straightforward fix being switching to
`create` which utilises the database.

Whilst there is a performance hit to using `create` this policy relies
on the `ActivityPolicy` which in turns loads objects from the database,
making stubbing pretty complicated.

For the moment, we will take the perfromance hit for working specs.