Export a CSV of all continuing activities
The first step in updating activities with the new transparency
identifier: producing a list of activities that fulfill the conditions
to continue, from the old GB-GOV-13 identifier corresponding to BEIS, to
the new GB-GOV-26 identifier corresponding to DSIT.

The list will need to be signed off by DSIT.

Usually we would have implemented this as a rake task to be run in the
server console, but we no longer have this level of access.

This ad-hoc export is temporary, and will be removed after all the
relevant activities are switched to the new identifier.