Download a list of non-continuing ODA activities
To help DSIT with confirming the lists of activities that will and won't
need to be updated with the new transparency identifier, allow them to
export a list of the activities that are assumed to not continue under
GB-GOV-26.

The reason we have used the criteria of elimination rather than a new
standalone query is that this is how the migration is likely to look
like in practice: activities that fulfill the continuation criteria will
be updated, and all other activities will be left as they are.

Non-ODA activities are excluded from both, because they are not
concerned with transparency reporting to begin with.