Change transparency identifier
For activities that will be continuing under DSIT:
- change their transparency identifier to replace the string "GB-GOV-13"
with "GB-GOV-26"; e.g. `GB-GOV-13-ISPF-UKSA-AH2TS7K-SPB9U87` becomes
`GB-GOV-26-ISPF-UKSA-AH2TS7K-SPB9U87`
- change the `providing_organisation_reference` for all their actuals,
from "GB-GOV-13" to "GB-GOV-26"
- change the `providing_organisation_reference` for all their forecasts,
from "GB-GOV-13" to "GB-GOV-26"

This type of data change would be better suited for a data migration or
a rake task, but both of those would require access to the application
console in the production environment.