Support feature flag to use new importer
The new Actuals, Refund and Activity comment importer is ready to be
tested. To allow us to do so we want to have control over which importer
the application uses.

Right now, we want to change as little as possible, keeping the existing
UI for testing.

Once we have confidence the importer is working as expected, we can come
back and make the most of the new behaviour.

The feature flag is managed by Rollout and is called:

`use_new_activity_actual_refund_comment_importer`