Retain compatibility with exiting import view
When extending the Import::RowError for use in the new importer we
changed the name of this attrbiute from `csv_row` to `csv_row_number` as
it was felt to be more descriptive.

However, whilst running both importers and switching between them, we
want to retain compatibility - the old importer uses `csv_row` so we
need to implement that for now, once we are confident in the new
importer we can strip out the old including this.