Exclude health check requests from host auth
Exclude health check requests from host authorisation, to allow the AWS
deploy pipeline to proceed faster. Currently health check requests are
denied by the host auth process, because the IPs are not in the DOMAIN
or ADDITIONAL_HOSTNAMES env vars.

We don't have a canonical list of the IPs, nor do we have easy access to
the env vars (these particular env vars live in task definitions, not
in ParameterStore).

https://api.rubyonrails.org/classes/ActionDispatch/HostAuthorization.html