Model a imported row
We want to help users by showing any rows that were skipped during a csv
import. To facilitate this we want a simple object that stores the row
number the imported item came from as well as the imported object
itself.

This is the same thing we do for rows that contain errors with the
Import::Error object.

We chose to namespace this to `csv` as the row number implies import
from at least tabulated data.