Ops: update 100 Activities' PO ID with data migration
See Zendesk 20045 in which we are requested to change the
`Activity#partner_organisation_identifier` for 100
activities.

These IDs are not used internally so there's no data
integrity issue within the RODA application.

We do however, log if:

- this change will result in there being another activity with
the same Partner Org ID and the same parent, as this is not
permitted ordinarily, see [Activity model][]

- the given Activity's current `partner_organisation_identifier`
does not match the ID listed in our spreadsheet of changes

Our data migration must be run with with either `--dry-run`
or `--live-run` e.g.

```sh
rails runner db/data/20240919000000_change_partner_org_ids.rb --dry-run`
```

In the `--dry-run` case we print out all the logging but
hold back on making the changes.

[Activity model]:
https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/blob/1c781b45d70571cce368da5d6ea414e79b6c7ba8/app/models/activity.rb#L154