Update and organise documentation
We want to try and leave a good set of documentation covering what we
learn in during the current working phase.

We'll take the time to update and organise the docs as a first step.

Our goal here is to offer at a glance to all the docs from the readme
with links directly to the detail.

We also deleted the unused `adr` directory.