Update docker compose for CI
We want to be clear about where this is usually used, CI so we rename
the file.

As this runs in CI there is no reason to map or bind volumes as the
entire runner is disposed of after each run.

We have to name `cached-from` the same as the build step in CI to make
the most of the caching.