Add new improved CI Github action
After extensive testing, this is the best (fastest) version we can get
to.

We use the Docker 'build and cache' action in a composite action so we
can share it across our jobs.

Each job loads the image layers from the Github action cache, updating
it as needed.

We ensure we set `RAILS_ENV` to test.

We create a job for each 'class' of test run in CI, this gives us some
time saving by being able to run the jobs in parallel.