Add hybrid_beis_dsit_activity flag
Some activities started under BEIS but will finish under DIST.

For these projects we want to include a IATI `other identifier` that
shows this by including the details of the 'previous reporting
organisation', i.e. BEIS is this case.

Rather than calculate this based on some tricky logic, we will instead
have a flag on Activity, we think this is better as:

- nothing from the data is actually output to the XML
- there will never be activities meeting this criteria now so the scope
  is fixed

Once we have the flag, we can migrate the correct set of activities and
display the additional XML.