Data migration
This data migration fixes the data in the application that was
previously updated in error.

We previously updated the `transparency_identifier` which should be
immutable. This data migration reverts that change.

We then set the new `hybrid_beis_dsit_activity` which makes the
application display 'previous reporting organisation' info in the IATI
xml.

A 'hybrid activity' is one that was started under BEIS but continued
under DSIT and so has to show:

- the previous reporting organisation
- retain the old IATI identifier